requirements.txt
# requirements.txt
openai>=1.3.0
argparse
requests>=2.0.0
Pillow>=9.0.0
google-generativeai>=0.3.0
